added nodechain generator and its test cases. removed load definition file and generate place holders from base config
