+++ src/service/config/configuration-service.ts
@@ -10,10 +10,12 @@ import { TreatmentOptions } from "@bc/domain/treatment-options";
 import { Node } from "@bc/domain/node";
 import { ProjectConfiguration } from "@bc/domain/configuration";
 import { FlowType } from "@bc/domain/inputs";
+import { NodeChainGenerator } from "@bc/service/config/nodechain-generator";
 
 @Service()
 export class ConfigurationService {
   private configuration: BaseConfiguration;
+  private _nodeChain: Node[];
 
   constructor() {
     switch (Container.get(constants.CONTAINER.ENTRY_POINT)) {
@@ -26,6 +28,11 @@ export class ConfigurationService {
       default:
         logAndThrow("Invalid entrypoint. Please contact with the administrator or report and issue to build-chain tool repository");
     }
+    this._nodeChain = [];
+  }
+
+  get nodeChain(): Node[] {
+    return this._nodeChain;
   }
 
   /**
@@ -33,6 +40,8 @@ export class ConfigurationService {
    */
   async init() {
     await this.configuration.init();
+    const nodeChainGenerator = new NodeChainGenerator(this.configuration);
+    this._nodeChain = await nodeChainGenerator.generateNodeChain(this.getStarterProjectName());
   }
 
   /**
@@ -61,7 +70,7 @@ export class ConfigurationService {
    * @returns {Node} starter node
    */
   getStarterNode(): Node {
-    const starterNode = this.configuration.projectList.find((node) => this.isNodeStarter(node));
+    const starterNode = this.nodeChain.find((node) => this.isNodeStarter(node));
     if (!starterNode) {
       logAndThrow(`There's no project ${this.getStarterProjectName()} in the chain
             This is normally due the project starting the job (or the one selected to behave like so it's not in the project tree information.
@@ -76,8 +85,8 @@ export class ConfigurationService {
    * @returns {NodeExecutionLevel} Upstream, current or downstream
    */
   getNodeExecutionLevel(node: Node): NodeExecutionLevel {
-    const starterNodeIndex = this.configuration.projectList.indexOf(this.getStarterNode());
-    const currentNodeIndex = this.configuration.projectList.indexOf(node);
+    const starterNodeIndex = this.nodeChain.indexOf(this.getStarterNode());
+    const currentNodeIndex = this.nodeChain.indexOf(node);
     if (currentNodeIndex < starterNodeIndex) {
       return NodeExecutionLevel.UPSTREAM;
     } else if (currentNodeIndex > starterNodeIndex) {
